requirements.txt
django
djangorestframework
rest_condition
